+++ extras/test/doubles/config_mock.h
@@ -67,7 +67,7 @@ class ConfigMock : public Supla::Config {
               (const char* key, const uint32_t value),
               (override));
   MOCK_METHOD(void, commit, (), (override));
-  MOCK_METHOD(void, saveWithDelay, (uint32_t delayMs), (override));
+  MOCK_METHOD(void, saveWithDelay, (uint16_t delayMs), (override));
   MOCK_METHOD(void, saveIfNeeded, (), (override));
   MOCK_METHOD(bool, generateGuidAndAuthkey, (), (override));
   MOCK_METHOD(bool, setDeviceName, (const char* name), (override));
